authentication/requirements.txt
Flask
mysql-connector
flask-mysql
flask_jwt